fix migration

+++ workspace/console/migrations/2020_07_29_182713_main_page_review_add_column_to_table.php
@@ -6,6 +6,7 @@ use Illuminate\Database\Schema\Blueprint;
 
 class MainPageReviewAddColumnToTable extends Migration
 {
+    private $table = 'main_page_review';
     /**
      * Run the migrations.
      *
@@ -13,10 +14,16 @@ class MainPageReviewAddColumnToTable extends Migration
      */
     public function up()
     {
-        App::$db->schema->table('main_page_review', function (Blueprint $table) {
-            $table->integer('tour_id')->unsigned()->default(0);
-            $table->foreign('tour_id')->references('id')->on('tour');
-            $table->integer('priority')->default(1);;
+        App::$db->schema->table($this->table, function (Blueprint $table) {
+            $is_tour_id_exist = App::$db->schema->hasColumn($this->table, 'tour_id');
+            if (!$is_tour_id_exist) {
+                // foreign key
+                $table->integer('tour_id')->unsigned()->default(0);
+            }
+            $is_priority_exist = App::$db->schema->hasColumn($this->table, 'priority');
+            if (!$is_priority_exist) {
+                $table->integer('priority')->default(1);
+            }
         });
     }
 
@@ -27,8 +34,9 @@ class MainPageReviewAddColumnToTable extends Migration
      */
     public function down()
     {
-        App::$db->schema->table('main_page_review', function (Blueprint $table) {
-            $table->dropColumn('tour_id');
+        App::$db->schema->table($this->table, function (Blueprint $table) {
+            if (App::$db->schema->hasColumn($this->table, 'tour_id')) $table->dropColumn('tour_id');
+            if (App::$db->schema->hasColumn($this->table, 'priority')) $table->dropColumn('tour_id');
         });
     }
 }
